+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/sync/SyncSession.kt
@@ -132,12 +132,14 @@ public interface SyncSession {
      * means that it is possible for one [Progress] instance to report
      * `isTransferComplete = true` and subsequent instances to report `isTransferComplete = false`.
      *
+     * The flow will be completed if the realm is closed.
+     *
      * The flow has an internal buffer of [Channel.BUFFERED] but if the consumer fails to consume the
      * elements in a timely manner the flow will be completed with an [IllegalStateException].
      *
      * @throws UnsupportedOperationException if invoked on a realm with Flexible Sync enabled.
      */
-    public fun progress(
+    public fun progressAsFlow(
         direction: Direction,
         progressMode: ProgressMode,
     ): Flow<Progress>
@@ -146,10 +148,12 @@ public interface SyncSession {
      * Create a [Flow] of [ConnectionStateChange]-events to receive notifications of updates to the
      * session's connection state.
      *
+     * The flow will be completed if the realm is closed.
+     *
      * The flow has an internal buffer of [Channel.BUFFERED] but if the consumer fails to consume
      * the elements in a timely manner the flow will be completed with an [IllegalStateException].
      */
-    public fun connectionState(): Flow<ConnectionStateChange>
+    public fun connectionStateAsFlow(): Flow<ConnectionStateChange>
 
     /**
      * Interface used to report any session errors.